requirements.txt
requests
pandas
python-dotenv
ShopifyAPI
google-cloud-bigquery
pyarrow
dbt-bigquery